requirements.txt
Amino.py == 1.2.10.1
websocket-client == 0.57.0